feat: add some tests & update type definition

+++ src/operators/pop/pop.ts
@@ -1,16 +1,19 @@
 import type { Collection } from '../../types';
 import { isArray, isCollection } from '../../utils';
 
-const collectionPopImpl = (collection: Collection): unknown => {
+const collectionPopImpl = (collection: Collection) => {
   if (isArray(collection)) {
     return collection.pop();
   }
 
   const result = [...collection].pop();
-  collection.delete([...collection.keys()].pop());
+  if (result) {
+    collection.delete([...collection.keys()].pop());
+  }
+
   return result;
 };
 
-export function pop() {
-  return <S>(source: S): unknown => (isCollection(source) ? collectionPopImpl(source) : source);
+export function pop(): <S extends Collection>(source: S) => unknown {
+  return <S extends Collection>(source: S): unknown => (isCollection(source) ? collectionPopImpl(source) : source);
 }
